chore: include ws281x source code in the project itself
This dependency has caused so much trouble that it is just easier to absorb it.  Building the jar file correctly with gradle has proven very tricky. Some of the errors encountered along the journey were:

- Security policy runtime error because java didn't like the way the jar was signed
- module not found error because gradle (for some reason) didn't include the module-info.class file
- class definition not found error because gradle didn't shade the outdated log4j dependency properly.

You get the picture. At this point, especially considering possible future API changes perhaps required due to performance or something similar, it is just way easier to absorb it rather than maintain it as a separate jar.